move stringFromTextAndArgs to utils, added test
