+++ infra/blueprint-test/pkg/bq/bq.go
@@ -118,7 +118,7 @@ func Run(t testing.TB, cmd string, opts ...cmdOption) gjson.Result {
 //
 // It fails the test if there are any errors executing the bq command or parsing the output value.
 func RunWithCmdOptsf(t testing.TB, opts []cmdOption, cmd string, args ...interface{}) gjson.Result {
-	return Run(t, stringFromTextAndArgs(append([]interface{}{cmd}, args...)...), opts...)
+	return Run(t, utils.StringFromTextAndArgs(append([]interface{}{cmd}, args...)...), opts...)
 }
 
 // Runf executes a bq command and returns value as gjson.Result.
@@ -127,39 +127,5 @@ func RunWithCmdOptsf(t testing.TB, opts []cmdOption, cmd string, args ...interfa
 //
 // It fails the test if there are any errors executing the bq command or parsing the output value.
 func Runf(t testing.TB, cmd string, args ...interface{}) gjson.Result {
-	return Run(t, stringFromTextAndArgs(append([]interface{}{cmd}, args...)...))
-}
-
-// stringFromTextAndArgs convert msg and args to formatted text
-func stringFromTextAndArgs(msgAndArgs ...interface{}) string {
-	if len(msgAndArgs) == 0 || msgAndArgs == nil {
-		return ""
-	}
-	if len(msgAndArgs) == 1 {
-		msg := msgAndArgs[0]
-		if msgAsStr, ok := msg.(string); ok {
-			return msgAsStr
-		}
-		return fmt.Sprintf("%+v", msg)
-	}
-	if len(msgAndArgs) > 1 {
-		return fmt.Sprintf(msgAndArgs[0].(string), msgAndArgs[1:]...)
-	}
-	return ""
-}
-
-// ActivateCredsAndEnvVars activates credentials and exports auth related envvars.
-func ActivateCredsAndEnvVars(t testing.TB, creds string) {
-	credsPath, err := utils.WriteTmpFile(creds)
-	if err != nil {
-		t.Fatal(err)
-	}
-	RunCmd(t, "auth activate-service-account", WithCommonArgs([]string{"--key-file", credsPath}))
-	// set auth related env vars
-	// TF provider auth
-	utils.SetEnv(t, "GOOGLE_CREDENTIALS", creds)
-	// bq SDK override
-	utils.SetEnv(t, "CLOUDSDK_AUTH_CREDENTIAL_FILE_OVERRIDE", credsPath)
-	// ADC
-	utils.SetEnv(t, "GOOGLE_APPLICATION_CREDENTIALS", credsPath)
+	return Run(t, utils.StringFromTextAndArgs(append([]interface{}{cmd}, args...)...))
 }
